fix seo metadata in client pages
