+++ app/layout.tsx
@@ -4,7 +4,25 @@ import Footer from '@/components/Footer';
 import FloatingMenu from '@/components/FloatingMenu';
 import FirebaseInit from '@/components/FirebaseInit';
 import Script from 'next/script';
+import type { Metadata } from 'next';
 import './globals.css';
+
+export const metadata: Metadata = {
+  metadataBase: new URL('https://labelretail.ci'),
+  title: {
+    default: 'Label Retail',
+    template: '%s | Label Retail',
+  },
+  description:
+    "Solutions de sécurité électronique et de télécommunication en Côte d'Ivoire",
+  alternates: { canonical: '/' },
+  openGraph: {
+    type: 'website',
+    locale: 'fr_FR',
+    url: 'https://labelretail.ci',
+    siteName: 'Label Retail',
+  },
+};
 
 export default function RootLayout({ children }: { children: ReactNode }) {
   return (
