+++ public/robots.txt
@@ -1,4 +1,4 @@
 User-agent: *
 Disallow:
 
-Sitemap: https://info@label-ci.com/sitemap.xml
+Sitemap: https://labelretail.ci/sitemap.xml
